It looks as if an explicit cast to Int64 is needed. Sounds like a compiler error, but it cures the problem.

+++ code/casa/implement/IO/LockFile.cc
@@ -241,7 +241,7 @@ void LockFile::getInfo (MemoryIO& info)
     if (leng > infoLeng) {
 	leng = infoLeng;
     }
-    info.seek (0);
+    info.seek (Int64(0));
     info.write (leng, buffer + SIZEREQID + SIZEINT);
     // Read the remaining info parts.
     if (infoLeng > leng) {
@@ -251,7 +251,7 @@ void LockFile::getInfo (MemoryIO& info)
 	info.write (infoLeng, buf);
 	delete [] buf;
     }
-    info.seek (0);
+    info.seek (Int64(0));
 }
 
 void LockFile::putInfo (const MemoryIO& info) const
